+++ rust/operator-binary/src/format/well_known.rs
@@ -1,5 +1,6 @@
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 use snafu::{OptionExt, Snafu};
+use stackable_operator::schemars::{self, JsonSchema};
 use strum::EnumDiscriminants;
 
 use super::{ConvertError, SecretFiles, convert};
@@ -16,14 +17,14 @@ const FILE_KERBEROS_KEYTAB_KRB5_CONF: &str = "krb5.conf";
 
 #[derive(Debug)]
 pub struct TlsPem {
-    pub certificate_pem: Vec<u8>,
-    pub key_pem: Vec<u8>,
+    pub certificate_pem: Option<Vec<u8>>,
+    pub key_pem: Option<Vec<u8>>,
     pub ca_pem: Vec<u8>,
 }
 
 #[derive(Debug)]
 pub struct TlsPkcs12 {
-    pub keystore: Vec<u8>,
+    pub keystore: Option<Vec<u8>>,
     pub truststore: Vec<u8>,
 }
 
@@ -36,7 +37,7 @@ pub struct Kerberos {
 #[derive(Debug, EnumDiscriminants)]
 #[strum_discriminants(
     name(SecretFormat),
-    derive(Deserialize),
+    derive(Serialize, Deserialize, JsonSchema),
     serde(rename_all = "kebab-case")
 )]
 pub enum WellKnownSecretData {
@@ -53,19 +54,23 @@ impl WellKnownSecretData {
                 key_pem,
                 ca_pem,
             }) => [
-                (names.tls_pem_cert_name, certificate_pem),
-                (names.tls_pem_key_name, key_pem),
-                (names.tls_pem_ca_name, ca_pem),
+                Some(names.tls_pem_cert_name).zip(certificate_pem),
+                Some(names.tls_pem_key_name).zip(key_pem),
+                Some((names.tls_pem_ca_name, ca_pem)),
             ]
-            .into(),
+            .into_iter()
+            .flatten()
+            .collect(),
             WellKnownSecretData::TlsPkcs12(TlsPkcs12 {
                 keystore,
                 truststore,
             }) => [
-                (names.tls_pkcs12_keystore_name, keystore),
-                (names.tls_pkcs12_truststore_name, truststore),
+                Some(names.tls_pkcs12_keystore_name).zip(keystore),
+                Some((names.tls_pkcs12_truststore_name, truststore)),
             ]
-            .into(),
+            .into_iter()
+            .flatten()
+            .collect(),
             WellKnownSecretData::Kerberos(Kerberos { keytab, krb5_conf }) => [
                 (FILE_KERBEROS_KEYTAB_KEYTAB.to_string(), keytab),
                 (FILE_KERBEROS_KEYTAB_KRB5_CONF.to_string(), krb5_conf),
@@ -84,13 +89,13 @@ impl WellKnownSecretData {
         if let Ok(certificate_pem) = take_file(SecretFormat::TlsPem, FILE_PEM_CERT_CERT) {
             let mut take_file = |file| take_file(SecretFormat::TlsPem, file);
             Ok(WellKnownSecretData::TlsPem(TlsPem {
-                certificate_pem,
-                key_pem: take_file(FILE_PEM_CERT_KEY)?,
+                certificate_pem: Some(certificate_pem),
+                key_pem: Some(take_file(FILE_PEM_CERT_KEY)?),
                 ca_pem: take_file(FILE_PEM_CERT_CA)?,
             }))
         } else if let Ok(keystore) = take_file(SecretFormat::TlsPkcs12, FILE_PKCS12_CERT_KEYSTORE) {
             Ok(WellKnownSecretData::TlsPkcs12(TlsPkcs12 {
-                keystore,
+                keystore: Some(keystore),
                 truststore: take_file(SecretFormat::TlsPkcs12, FILE_PKCS12_CERT_TRUSTSTORE)?,
             }))
         } else if let Ok(keytab) = take_file(SecretFormat::Kerberos, FILE_KERBEROS_KEYTAB_KEYTAB) {
@@ -181,6 +186,18 @@ pub struct NamingOptions {
         default = "default_tls_pem_ca_name"
     )]
     pub tls_pem_ca_name: String,
+}
+
+impl Default for NamingOptions {
+    fn default() -> Self {
+        Self {
+            tls_pkcs12_keystore_name: default_pkcs12_keystore_name(),
+            tls_pkcs12_truststore_name: default_pkcs12_truststore_name(),
+            tls_pem_cert_name: default_tls_pem_cert_name(),
+            tls_pem_key_name: default_tls_pem_key_name(),
+            tls_pem_ca_name: default_tls_pem_ca_name(),
+        }
+    }
 }
 
 fn default_pkcs12_keystore_name() -> String {